refactor: controller refactored
